+++ qiime/alpha_diversity.py
@@ -6,7 +6,7 @@ __credits__ = ["Justin Kuczynski", "Rob Knight", "Greg Caporaso",
                "William Van Treuren", "Jose Antonio Navas Molina",
                "Jai Ram Rideout"]
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "Justin Kuczynski"
 __email__ = "justinak@gmail.com"