+++ qiime/identify_chimeric_seqs.py
@@ -28,7 +28,7 @@ __copyright__ = "Copyright 2011, The QIIME Project"
 __credits__ = ["Greg Caporaso", "Jens Reeder", "William Walters",
                "Jai Ram Rideout", "Adam Robbins-Pianka"]
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "Greg Caporaso"
 __email__ = "gregcaporaso@gmail.com"