+++ qiime/join_paired_ends.py
@@ -4,7 +4,7 @@ __author__ = "Mike Robeson"
 __copyright__ = "Copyright 2013, The QIIME Project"
 __credits__ = ["Mike Robeson"]
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "Mike Robeson"
 __email__ = "robesonms@ornl.gov"