+++ scripts/identify_paired_differences.py
@@ -6,7 +6,7 @@ __author__ = "Greg Caporaso"
 __copyright__ = "Copyright 2013, The QIIME project"
 __credits__ = ["Greg Caporaso", "Jose Carlos Clemente Litran"]
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "Greg Caporaso"
 __email__ = "gregcaporaso@gmail.com"