+++ scripts/parallel_pick_otus_sortmerna.py
@@ -7,7 +7,7 @@ Parallel launching script for SortMeRNA v 2.0
 
 from __future__ import division
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "Jenya Kopylov"
 __email__ = "jenya.kopylov@gmail.com"