+++ scripts/split_libraries.py
@@ -9,7 +9,7 @@ __credits__ = ["Rob Knight", "Micah Hamady", "Greg Caporaso",
                "Jens Reeder", "Jose Antonio Navas Molina",
                "Jai Ram Rideout"]  # remember to add yourself
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "William Walters"
 __email__ = "william.a.walters@colorado.edu"