+++ tests/test_make_library_id_lists.py
@@ -8,7 +8,7 @@ __copyright__ = "Copyright 2011, The QIIME Project"
 # remember to add yourself if you make changes
 __credits__ = ["Rob Knight"]
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "Kyle Bittinger"
 __email__ = "kylebittinger@gmail.com"