+++ tests/test_stats.py
@@ -7,7 +7,7 @@ __credits__ = ["Jai Ram Rideout", "Michael Dwan", "Logan Knecht",
                "Damien Coy", "Levi McCracken", "Andrew Cochran",
                "Will Van Treuren"]
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "Jai Ram Rideout"
 __email__ = "jai.rideout@gmail.com"