+++ tests/test_supervised_learning.py
@@ -7,7 +7,7 @@ __copyright__ = "Copyright 2011, The QIIME Project"
 # remember to add yourself if you make changes
 __credits__ = ["Dan Knights", "Luke Ursell"]
 __license__ = "GPL"
-__version__ = "1.8.0-dev"
+__version__ = "1.9.0-rc1"
 __maintainer__ = "Dan Knights"
 __email__ = "daniel.knights@colorado.edu"